requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
anyio==4.12.1
astroid==4.0.4
attrs==25.4.0
black==26.1.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
coverage==7.13.4
cycler==0.12.1
dill==0.4.1
filelock==3.21.2
flake8==7.3.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mccabe==0.7.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.5.1
pluggy==1.6.0
polars==1.38.1
polars-runtime-32==1.38.1
propcache==0.4.1
psutil==7.2.2
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.11.0
pylint==4.0.4
pyparsing==3.3.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-telegram-bot==22.6
pytokens==0.4.1
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.0
setuptools==82.0.0
six==1.17.0
sympy==1.14.0
tomlkit==0.14.0
torch==2.10.0
torchvision==0.25.0
twilio==9.10.1
typing_extensions==4.15.0
ultralytics==8.4.14
ultralytics-thop==2.0.18
urllib3==2.6.3
yarl==1.22.0
